refactor(sha256): collapse the two copies into shared/sha256
projects/cuid carried a copy of rocprofiler-sdk's SHA-256, kept in step by a
drift script and a CI job.

It was a copy because cuid sits below rocprofiler-sdk and cannot link it. That
applies to linking a library; SHA-256 depends on nothing but the standard
library, so as a leaf built from source it adds no edge either way.

Both consumers now add_subdirectory() shared/sha256, as libhsakmt does for
wkmi. lib/common/sha256.hpp becomes a shim, so its callers are unchanged. The
drift check and ROCPROFILER_SDK_SHA256_STANDALONE go with it.

Nine workflows check out only projects/rocprofiler-sdk, so the new dependency
on shared/sha256 has to be added to each of their sparse-checkout lists or they
fail at configure time against a checkout that cannot satisfy the guard.
