+++ .github/workflows/hipfile-profiler-abi.yml
@@ -47,6 +47,7 @@ jobs:
           sparse-checkout: |
             projects/hipfile
             projects/rocprofiler-sdk
+            shared/sha256
             .github/workflows
       - name: Authenticating to GitHub Container Registry
         uses: docker/login-action@650006c6eb7dba73a995cc03b0b2d7f5ca915bee # v4.2.0
