+++ .github/workflows/rocprof-trace-decoder-continuous_integration.yml
@@ -175,6 +175,7 @@ jobs:
           sparse-checkout: |
             .github/actions/rocprofiler-sdk-util
             projects/rocprofiler-sdk
+            shared/sha256
             projects/aqlprofile
             projects/rocprofiler-register
             projects/rocprof-trace-decoder
