+++ .github/workflows/rocprofiler-sdk-code_coverage.yml
@@ -135,6 +135,7 @@ jobs:
         sparse-checkout: |
           .github/actions/rocprofiler-sdk-util
           projects/rocprofiler-sdk
+          shared/sha256
           projects/aqlprofile
           projects/rocprofiler-register
           projects/rocr-runtime
