+++ .github/workflows/rocprofiler-sdk-continuous_integration.yml
@@ -150,6 +150,7 @@ jobs:
           sparse-checkout: |
             .github/actions/rocprofiler-sdk-util
             projects/rocprofiler-sdk
+            shared/sha256
             projects/aqlprofile
             projects/rocprofiler-register
             projects/rocprof-trace-decoder
@@ -383,6 +384,7 @@ jobs:
           sparse-checkout: |
             .github/actions/rocprofiler-sdk-util
             projects/rocprofiler-sdk
+            shared/sha256
             projects/aqlprofile
             projects/rocprofiler-register
             projects/rocprof-trace-decoder
@@ -529,6 +531,7 @@ jobs:
         sparse-checkout: |
           .github/actions/rocprofiler-sdk-util
           projects/rocprofiler-sdk
+          shared/sha256
           projects/aqlprofile
           projects/rocprofiler-register
           projects/rocprof-trace-decoder
