+++ .github/workflows/rocprofiler-sdk-docs.yml
@@ -75,7 +75,9 @@ jobs:
       - name: Checkout
         uses: actions/checkout@v6
         with:
-          sparse-checkout: projects/rocprofiler-sdk
+          sparse-checkout: |
+            projects/rocprofiler-sdk
+            shared/sha256
           submodules: false
           set-safe-directory: true
       - name: Init/Update submodules
@@ -143,6 +145,7 @@ jobs:
           sparse-checkout: |
             .github/actions/rocprofiler-sdk-util
             projects/rocprofiler-sdk
+            shared/sha256
             projects/rocr-runtime
             projects/clr
             projects/hip
