+++ .github/workflows/rocprofiler-sdk-formatting.yml
@@ -33,7 +33,9 @@ jobs:
     steps:
     - uses: actions/checkout@v6
       with:
-        sparse-checkout: projects/rocprofiler-sdk
+        sparse-checkout: |
+          projects/rocprofiler-sdk
+          shared/sha256
 
     - name: Extract branch name
       shell: bash
@@ -73,7 +75,9 @@ jobs:
     steps:
     - uses: actions/checkout@v6
       with:
-        sparse-checkout: projects/rocprofiler-sdk
+        sparse-checkout: |
+          projects/rocprofiler-sdk
+          shared/sha256
 
     - name: Extract branch name
       shell: bash
@@ -120,7 +124,9 @@ jobs:
     steps:
     - uses: actions/checkout@v6
       with:
-        sparse-checkout: projects/rocprofiler-sdk
+        sparse-checkout: |
+          projects/rocprofiler-sdk
+          shared/sha256
 
     - name: Extract branch name
       shell: bash
@@ -158,7 +164,9 @@ jobs:
     steps:
     - uses: actions/checkout@v6
       with:
-        sparse-checkout: projects/rocprofiler-sdk
+        sparse-checkout: |
+          projects/rocprofiler-sdk
+          shared/sha256
 
     - name: Extract branch name
       shell: bash
@@ -185,7 +193,9 @@ jobs:
     steps:
     - uses: actions/checkout@v6
       with:
-        sparse-checkout: projects/rocprofiler-sdk
+        sparse-checkout: |
+          projects/rocprofiler-sdk
+          shared/sha256
 
     - name: Set up Python
       uses: actions/setup-python@v5
