+++ .github/workflows/rocprofiler-sdk-python.yml
@@ -46,7 +46,9 @@ jobs:
     steps:
     - uses: actions/checkout@v6
       with:
-        sparse-checkout: projects/rocprofiler-sdk
+        sparse-checkout: |
+          projects/rocprofiler-sdk
+          shared/sha256
     - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v5
       with:
