+++ projects/cuid/lib/CMakeLists.txt
@@ -34,7 +34,6 @@ set(CUID_SOURCES
     ${SRC_DIR}/cuid_platform.cc
     ${SRC_DIR}/cuid_file.cc
     ${SRC_DIR}/hmac.cc
-    ${SRC_DIR}/sha256.cc
     ${SRC_DIR}/pci_util.cc
     ${SRC_DIR}/acpi_parser.cc
     ${SRC_DIR}/smbios_util.cc
@@ -53,7 +52,6 @@ set(CUID_HEADERS
     ${SRC_DIR}/cuid_platform.h
     ${SRC_DIR}/cuid_file.h
     ${SRC_DIR}/hmac.h
-    ${SRC_DIR}/sha256.h
     ${SRC_DIR}/ipc_protocol.h
     ${SRC_DIR}/pci_util.h
     ${SRC_DIR}/acpi_parser.h
@@ -73,8 +71,25 @@ target_include_directories(amdcuid_static PUBLIC ${CMAKE_CURRENT_SOURCE_DIR} PRI
 # the HMAC key is shared and CUIDs remain stable across version upgrades.
 target_compile_definitions(amdcuid_static PRIVATE AMDCUID_CONFIG_DIR="${AMDCUID_SHARED_CONFIG_DIR}")
 
-# SHA-256/HMAC comes from src/sha256.cc, so the only external dependency left
-# is the platform CSPRNG: bcrypt on Windows, getrandom(2)//dev/urandom on POSIX.
+# SHA-256/HMAC comes from shared/sha256, built from source.
+set(ROCM_SHA256_DIR
+    "${CMAKE_CURRENT_SOURCE_DIR}/../../../shared/sha256"
+    CACHE PATH
+    "Path to the shared SHA-256 component"
+)
+if(NOT TARGET rocm-sha256)
+    if(NOT EXISTS "${ROCM_SHA256_DIR}/CMakeLists.txt")
+        message(
+            FATAL_ERROR
+            "Shared SHA-256 not found at ${ROCM_SHA256_DIR}. Set ROCM_SHA256_DIR "
+            "when building cuid outside the rocm-systems tree."
+        )
+    endif()
+    add_subdirectory(${ROCM_SHA256_DIR} ${CMAKE_CURRENT_BINARY_DIR}/rocm-sha256)
+endif()
+target_link_libraries(amdcuid_static PUBLIC rocm-sha256)
+
+# The only external dependency left is the platform CSPRNG.
 if(WIN32)
     target_link_libraries(amdcuid_static PUBLIC bcrypt)
 endif()
